consolidate libwayshot api
