+++ libwayshot/src/image_util.rs
