+++ libwayshot/src/screencopy.rs
@@ -1,0 +1,136 @@
+use std::{
+    ffi::CStr,
+    os::unix::prelude::RawFd,
+    time::{SystemTime, UNIX_EPOCH},
+};
+
+use anyhow::{anyhow, bail, Result};
+use image::{ColorType, DynamicImage, ImageBuffer, Pixel};
+use memmap2::MmapMut;
+use nix::{
+    fcntl,
+    sys::{memfd, mman, stat},
+    unistd,
+};
+use wayland_client::protocol::{wl_output, wl_shm::Format};
+
+/// Type of frame supported by the compositor. For now we only support Argb8888, Xrgb8888, and
+/// Xbgr8888.
+#[derive(Debug, Copy, Clone, PartialEq)]
+pub struct FrameFormat {
+    pub format: Format,
+    pub width: u32,
+    pub height: u32,
+    pub stride: u32,
+}
+
+fn create_image_buffer<P>(
+    frame_format: &FrameFormat,
+    frame_mmap: &MmapMut,
+) -> Result<ImageBuffer<P, Vec<P::Subpixel>>>
+where
+    P: Pixel<Subpixel = u8>,
+{
+    ImageBuffer::from_vec(frame_format.width, frame_format.height, frame_mmap.to_vec())
+        .ok_or(anyhow!("buffer too large"))
+}
+
+/// The copied frame comprising of the FrameFormat, ColorType (Rgba8), and a memory backed shm
+/// file that holds the image data in it.
+#[derive(Debug)]
+pub struct FrameCopy {
+    pub frame_format: FrameFormat,
+    pub frame_color_type: ColorType,
+    pub frame_mmap: MmapMut,
+    pub transform: wl_output::Transform,
+}
+
+impl TryFrom<&FrameCopy> for DynamicImage {
+    type Error = anyhow::Error;
+
+    fn try_from(value: &FrameCopy) -> Result<Self> {
+        Ok(match value.frame_color_type {
+            ColorType::Rgb8 => DynamicImage::ImageRgb8(create_image_buffer(
+                &value.frame_format,
+                &value.frame_mmap,
+            )?),
+            ColorType::Rgba8 => DynamicImage::ImageRgba8(create_image_buffer(
+                &value.frame_format,
+                &value.frame_mmap,
+            )?),
+            _ => bail!("invalid image color type"),
+        })
+    }
+}
+
+/// Return a RawFd to a shm file. We use memfd create on linux and shm_open for BSD support.
+/// You don't need to mess around with this function, it is only used by
+/// capture_output_frame.
+pub fn create_shm_fd() -> std::io::Result<RawFd> {
+    // Only try memfd on linux and freebsd.
+    #[cfg(any(target_os = "linux", target_os = "freebsd"))]
+    loop {
+        // Create a file that closes on succesful execution and seal it's operations.
+        match memfd::memfd_create(
+            CStr::from_bytes_with_nul(b"libwayshot\0").unwrap(),
+            memfd::MemFdCreateFlag::MFD_CLOEXEC | memfd::MemFdCreateFlag::MFD_ALLOW_SEALING,
+        ) {
+            Ok(fd) => {
+                // This is only an optimization, so ignore errors.
+                // F_SEAL_SRHINK = File cannot be reduced in size.
+                // F_SEAL_SEAL = Prevent further calls to fcntl().
+                let _ = fcntl::fcntl(
+                    fd,
+                    fcntl::F_ADD_SEALS(
+                        fcntl::SealFlag::F_SEAL_SHRINK | fcntl::SealFlag::F_SEAL_SEAL,
+                    ),
+                );
+                return Ok(fd);
+            }
+            Err(nix::errno::Errno::EINTR) => continue,
+            Err(nix::errno::Errno::ENOSYS) => break,
+            Err(errno) => return Err(std::io::Error::from(errno)),
+        }
+    }
+
+    // Fallback to using shm_open.
+    let sys_time = SystemTime::now();
+    let mut mem_file_handle = format!(
+        "/libwayshot-{}",
+        sys_time.duration_since(UNIX_EPOCH).unwrap().subsec_nanos()
+    );
+    loop {
+        match mman::shm_open(
+            // O_CREAT = Create file if does not exist.
+            // O_EXCL = Error if create and file exists.
+            // O_RDWR = Open for reading and writing.
+            // O_CLOEXEC = Close on succesful execution.
+            // S_IRUSR = Set user read permission bit .
+            // S_IWUSR = Set user write permission bit.
+            mem_file_handle.as_str(),
+            fcntl::OFlag::O_CREAT
+                | fcntl::OFlag::O_EXCL
+                | fcntl::OFlag::O_RDWR
+                | fcntl::OFlag::O_CLOEXEC,
+            stat::Mode::S_IRUSR | stat::Mode::S_IWUSR,
+        ) {
+            Ok(fd) => match mman::shm_unlink(mem_file_handle.as_str()) {
+                Ok(_) => return Ok(fd),
+                Err(errno) => match unistd::close(fd) {
+                    Ok(_) => return Err(std::io::Error::from(errno)),
+                    Err(errno) => return Err(std::io::Error::from(errno)),
+                },
+            },
+            Err(nix::errno::Errno::EEXIST) => {
+                // If a file with that handle exists then change the handle
+                mem_file_handle = format!(
+                    "/libwayshot-{}",
+                    sys_time.duration_since(UNIX_EPOCH).unwrap().subsec_nanos()
+                );
+                continue;
+            }
+            Err(nix::errno::Errno::EINTR) => continue,
+            Err(errno) => return Err(std::io::Error::from(errno)),
+        }
+    }
+}
