+++ wayshot/Cargo.toml
@@ -19,12 +19,9 @@ clap = "4.3.11"
 env_logger = { version = "0.10.0", default-features = false, features = ["auto-color", "color"] }
 log = "0.4.19"
 
-wayland-client = "0.30.2"
-wayland-protocols = { version = "0.30.0", features=["client", "unstable"] }
-
 libwayshot = { version="0.1.2", path = "../libwayshot" }
 
-image = { version = "0.24", default-features = false, features = ["jpeg", "png", "pnm"] }
+image = { version = "0.24", default-features = false, features = ["jpeg", "png", "pnm", "qoi"] }
 
 dialoguer = { version = "0.10.4", features = ["fuzzy-select"] }
